+++ src/main/java/ua/com/reactive/uss_yaroslav/entity/Ingredient.java
@@ -8,12 +8,12 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 public class Ingredient {
-    private Long id;
+    private int id;
     private String name;
     private double price;
     private int quantity;
 
-    public Ingredient(Long id, String name, double price, int quantity) {
+    public Ingredient(int id, String name, double price, int quantity) {
         this.id = id;
         this.name = name;
         this.price = price;
